+++ catalyst/registry/tests/registery_foo.py
