+++ catalyst/tools/__init__.py
@@ -1,5 +1,4 @@
 # flake8: noqa
 from .frozen_class import FrozenClass
-from .registry import Registry, RegistryException
 from .settings import settings
 from .time_manager import TimeManager
